fix(user-updates) updated users load for admin page only
